refactor(src): 重构路径引用并重命名部分模块
- 统一调整内部模块导入路径，移除多余的 system 子目录
- 重命名 common/runtime/state.ts 为 common/state.ts 并同步更新相关引用
- 重命名 common/runtime/validate.ts 为 common/validate.ts 并同步更新相关引用
- 重命名 common/runtime/file-history.ts 为 common/file-history 并同步更新相关引用
- 更新测试文件和工具模块中对应的导入路径，确保一致性
- 保持代码逻辑不变，仅调整代码结构和模块路径优化维护性
