+++ src/common/bash-timeout.ts
