+++ src/tests/shell-utils.test.ts
@@ -7,8 +7,8 @@ import {
   resolveWindowsGitBashPath,
   rewriteWindowsNullRedirect,
   windowsPathToPosixPath,
-} from "../common/system/shell-utils";
-import { isAbsoluteFilePath, normalizeFilePath } from "../common/runtime/state";
+} from "../common/shell-utils";
+import { isAbsoluteFilePath, normalizeFilePath } from "../common/state";
 
 test("Windows paths convert to Git Bash POSIX paths", () => {
   assert.equal(windowsPathToPosixPath("C:\\Users\\foo"), "/c/Users/foo");
