+++ src/ui/views/ProcessStdoutView.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useMemo, useRef, useState } from "react";
 import { Box, Text } from "ink";
-import { BASH_TIMEOUT_DECREMENT_MS, BASH_TIMEOUT_INCREMENT_MS } from "../../common/system/bash-timeout";
+import { BASH_TIMEOUT_DECREMENT_MS, BASH_TIMEOUT_INCREMENT_MS } from "../../common/bash-timeout";
 import type { BashTimeoutAdjustment, SessionEntry, SessionProcessEntry } from "../../session";
 import { useTerminalInput } from "../hooks";
